requirements.txt
discord.py
python-i18n
psycopg2
